Changed *Extension classes to *Helper classes.
The change was made because I feel it is not a good practice to create TypeExtensions. Type extensions are less discoverable and give the impression that the type has a magical new fonction. Helper classes are easier to spot in the code and denote a clear intention from the pa of the programmer.